Work on tests continued
